move timer module into root directory
