+++ pyproject.toml
@@ -49,7 +49,6 @@ lint = [
 
 [tool.setuptools]
 packages = ["nominatim_data_analyser",
-            "nominatim_data_analyser.logger",
             "nominatim_data_analyser.core",
             "nominatim_data_analyser.core.exceptions",
             "nominatim_data_analyser.core.yaml_logic",
