+++ src/nominatim_data_analyser/timer.py
